backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.0
openai==1.12.0
azure-communication-email==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.10
pytest==7.4.3
httpx==0.26.0
email-validator==2.2.0
